separate the sources into dedicated directory, fix doc generation
for the documentation generation to work, the code needs to work
as a package, and the package needs to have a valid name
(kyber-py won't work), so put all of the sources into the src/kyber
directory. Upside: clear separation between shipping code and tests.
Also, because everything is in single module, it won't pollute
python namespace

Signed-off-by: Hubert Kario <hkario@redhat.com>

# Conflicts:
#	src/kyber_py/kyber/kyber.py
#	src/kyber_py/ml_kem/ml_kem.py